requirements.txt
APScheduler==3.6.3
attrs==22.2.0
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
clarifai-grpc==8.10.0
dnspython==2.2.1
emoji==2.2.0
ephem==4.1.3
exceptiongroup==1.0.4
flake8==6.0.0
googleapis-common-protos==1.56.4
grpcio==1.51.1
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
mongomock==4.1.2
packaging==22.0
pluggy==1.0.0
protobuf==3.19.3
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.3.3
PySocks==1.7.1
pytest==7.2.0
pytest-mongodb==2.2.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
sentinels==1.0.0
six==1.16.0
tomli==2.0.1
tornado==6.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.13
